Change Action to inherit Category in Core
The concept of Action representing the "invocable operation" itself is
removed. Instead an Action is just an *identifier* of the operation.

An Action instance identifies an invocable operation in much the same
way as a Kind instance identifies an Entity sub-type.

Impact on the existing HTTP Rendering (occi/1.1) is none. In fact the
text/ renderings already use "type=action" in its Category headers.