Protocol data-format split, work in progress
The purpose of the http_protocol document is to describe the HTTP
protocol part of OCCI without involving the data-format of the payload.
This will make other data-formats such as JSON easy to incorporate into
the OCCI specification.

The document describes what each HTTP verb does for element URIs and
collection URIs respectively. This way it should be straight forward
for both client and server implementors to find out exactly what each
HTTP verb is supposed to do.